third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-unprefixed-after-prefixed-expected.txt


EVENT(loadstart)
Call unprefixed EME API after prefixed EME API.
RUN(video.webkitGenerateKeyRequest('webkit-org.w3.clearkey'))
RUN(mediaKeys = new MediaKeys("org.w3.clearkey"))
EXPECTED (mediaKeys != 'null') OK
TEST(video.setMediaKeys(mediaKeys)) THROWS(DOMException.INVALID_STATE_ERR: Failed to execute 'setMediaKeys' on 'HTMLMediaElement': Mixed use of EME prefixed and unprefixed API not allowed.) OK
END OF TEST
